CAFFIENE_CLAUSE__SENTINAL_GEC/ai-theft-detection/requirements.txt
opencv-python>=4.7.0
numpy>=1.19.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
requests>=2.31.0
ultralytics>=8.0.0
pandas>=1.3.0
flask>=2.0.0
werkzeug>=2.0.0
tqdm>=4.64.0
jinja2>=3.1.0
markupsafe>=2.1.0
itsdangerous>=2.1.0
click>=8.1.0
joblib>=1.1.0
mediapipe==0.10.5
xgboost>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.12.0
beautifulsoup4>=4.11.0
zipfile36>=0.1.0